Finished initial version of firmware and python interface.

+++ firmware/constants.cpp
@@ -1,10 +1,13 @@
 #include "constants.h"
 
 const uint32_t Baudrate = 115200; 
-const uint32_t Number_Of_Pwm = 24;
-const uint16_t Number_Of_Device = 1;
+
+const int32_t Number_Of_Pwm = 24;
+const int32_t Number_Of_Device = 1;
+
 const uint16_t Max_Pwm_Value = 4095;
 const uint16_t Min_Pwm_Value = 0;
+
 const uint8_t Enable_Pin = 9;
 const uint8_t Latch_Pin = 10;
 const uint8_t Data_Pin = 11;
